requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-storage
google-cloud-pubsub
supervision==0.27.0
inference-gpu==0.9.13
transformers
opencv-python-headless
torch
numpy
# Roboflow Sports repo from the notebook
git+https://github.com/roboflow/sports.git@feat/basketball